[Spark] fix missing BigQuery class in column lineage
